Refactor + use int2bool for domain encoding
