+++ tests/test_lazy_gurobi.py
@@ -240,7 +240,7 @@ def get_envs():
 
         yield e | {"solver_kwargs": {"env": debug_env}}
 
-    for encoding in [Encoding.DEFAULT, Encoding.GLEB, Encoding.XCSP3, Encoding.BOOL_GLEB]:
+    for encoding in Encoding:
         yield {
             "alias": f"base_gurobi-{encoding}",
             "solver": CPM_gurobi,
